+++ lib/chef/resource/powershell_script.rb
@@ -15,17 +15,39 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-
 require 'chef/resource/windows_script'
 
 class Chef
   class Resource
     class PowershellScript < Chef::Resource::WindowsScript
 
+      set_guard_inherited_attributes(:architecture)
+
       def initialize(name, run_context=nil)
         super(name, run_context, :powershell_script, "powershell.exe")
+        @convert_boolean_return = false
       end
 
+      def convert_boolean_return(arg=nil)
+        set_or_return(
+          :convert_boolean_return,
+          arg,
+          :kind_of => [ FalseClass, TrueClass ]
+        )
+      end
+
+      protected
+
+      # Allow callers evaluating guards to request default
+      # attribute values. This is needed to allow
+      # convert_boolean_return to be true in guard context by default,
+      # and false by default otherwise. When this mode becomes the
+      # default for this resource, this method can be removed since
+      # guard context and recipe resource context will have the
+      # same behavior.
+      def self.get_default_attributes(opts)
+        {:convert_boolean_return => true}
+      end
     end
   end
 end