+++ CHANGELOG.md
@@ -65,6 +65,7 @@
 * Increase bootstrap log_level when knife -V -V is set (CHEF-3610)
 * Knife cookbook test should honor chefignore (CHEF-4203)
 * Fix ImmutableMash and ImmutableArray to_hash and to_a methods (CHEF-5132)
+* guard_interpreter attribute: use powershell\_script, other script resources in guards (CHEF-4553)
 
 ## Last Release: 11.10.0 (02/06/2014)